+++ lib/generators/cms/component/social_sharing/templates/app/cells/social_sharing_cell.rb
@@ -1,25 +1,35 @@
 class SocialSharingCell < BoxCell
+  SUPPORTED_PROVIDER = [:facebook, :google, :twitter, :linkedin]
+
   # Cell actions:
 
-  def show(url = nil, provider = [])
+  def show(url = nil, provider = SUPPORTED_PROVIDER)
     @url = url
-
-    if provider.present?
-      @provider = provider.find_all{|item| available_provider.include?(item) }
-    else
-      @provider = available_provider
-    end
+    @provider = provider
 
     render
   end
 
-  #Cell states:
+  # Cell states:
+  # The following states assume @url to be given.
 
   def facebook
+    @locale = 'en_US'
+    @data = {
+      href: @url,
+      send: 'false',
+      layout: 'button_count',
+      show: {
+        faces: 'false',
+      },
+    }
+
     render
   end
 
   def google
+    @locale = 'en_US'
+
     render
   end
 
@@ -29,11 +39,5 @@ class SocialSharingCell < BoxCell
 
   def linkedin
     render
-  end
-
-  private
-
-  def available_provider
-    @valid_provider ||= [:facebook, :google, :twitter, :linkedin]
   end
 end